requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
outcome==1.2.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
trio==0.22.0
trio-websocket==0.9.2
twilio==7.16.4
urllib3==1.26.14
wsproto==1.2.0
